+++ native/common/include/jp_classloader.h
@@ -12,7 +12,7 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
-   
+
  *****************************************************************************/
 #ifndef _JPCLASSLOADER_H_
 #define _JPCLASSLOADER_H_