+++ native/common/jp_floattype.cpp
@@ -85,7 +85,7 @@ JPMatch::Type JPFloatType::getJavaConversion(JPMatch& match, JPJavaFrame& frame,
 	if (JPPyObject::isNone(pyobj))
 		return JPMatch::_none;
 
-	JPValue* value = JPPythonEnv::getJavaValue(pyobj);
+	JPValue *value = JPPythonEnv::getJavaValue(pyobj);
 	if (value != NULL)
 	{
 		if (value->getClass() == this)