+++ doc/CHANGELOG.rst
@@ -16,6 +16,25 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
     with older versions requires use of the IPython greedy method.
 
 
+  - Updated getJVMVersion to work with JDK 9+.
+
+  - Fixed an issue with JDK 12 regarding calling methods with reflection.
+  
+  - Removed limitations having to do with CallerSensitive methods. Methods
+    affected are listed in :doc:`caller-sensitive`. Caller sensitive 
+    methods now receive an internal JPype class as the desut
+
+  - Fixed incorrect change with exceptions ``str()``
+  
+  - Fixed segfault when converting null elements while accessing a slice
+    from a Java object array.
+
+  - JProxy objects now are returned from Java as the Python objects 
+    that originate from. Older style proxy classes return the 
+    inst or dict. New style return the proxy class instance.
+    Thus proxy classes can be stored on generic Java containers
+    and retrieved as Python objects.
+
 - **0.7.0 - 2019**
 
   - Doc strings are generated for classes and methods.